added functionality to visualize, add and delete products for all authenticated users + visualize, add and delete users for admin and developer roles + fixed bad input and other edge case bugs

+++ demo/src/main/java/ro/utcluj/demo/dto/ProductDto.java
@@ -4,5 +4,5 @@ import lombok.Builder;
 import ro.utcluj.demo.model.Category;
 
 @Builder
-public record ProductDto(Long id, String name, Category category, Long stock, Long price) {
+public record ProductDto(Integer id, String name, Category category, Long stock, Long price) {
 }
